Build & Test Success On Linux | Refactored Build Scripts to account for LTO & Bugs in linux | Added Proper Floating Point comparisons to the C++ implementation using googletests' FloatingPoint Type | Turned on Aggressive compiler flags that I forgot to turn on - refactored code with that in mind | Improved CMake scripts by modularizing to simple files and using accordingly in the top-level CMakeLists.txt | Updated to latest commit in util2

+++ tree/include/tree/internal/nodeMetadata.hpp
@@ -29,8 +29,8 @@ public:
 
     uint32_t getConditionalChild(bool trueLeftFalseRight) const noexcept {
         const uint64_t bitMask  = trueLeftFalseRight ? k_leftChildBitmask  : k_rightChildBitmask;
-        const uint32_t bitShift = trueLeftFalseRight ? k_leftChildBitShift : k_rightChildBitShift;
-        return (m_data & bitMask) >> bitShift;
+        const uint64_t bitShift = trueLeftFalseRight ? k_leftChildBitShift : k_rightChildBitShift;
+        return static_cast<uint32_t>( (m_data & bitMask) >> bitShift );
     }
 
     uint32_t getLeftChild() const noexcept {
